chore: clean compiler warnings, update docs with real binary metrics
- Add #![allow(dead_code)] to platform-specific linux.rs modules (DiskInfo/
  NetInfo inline buffers and public accessors are part of the self-contained
  struct layout, not read directly)
- Add #[allow(unused_imports)] to per-platform MetricErr/Info re-exports
  that are part of the public API surface but unused within the binary
- Remove dead 'padded' tracking in src/tls.rs base64 decoder (the trimmed
  'end' index already drives output sizing)
- Update docs (README, spec, MASTER progress) with measured binary metrics

Verification:
- Windows dev build: 0 warnings, 162 tests pass
- WSL musl release build: 0 warnings, 174 tests pass
- Binary: 1.5M (x86_64-unknown-linux-musl)

+++ src/monitor/ip/macos.rs
@@ -40,8 +40,8 @@ impl From<std::io::Error> for MetricErr {
 // ═══════════════════════════════════════════════════════════════════
 
 // macOS uses AF_INET=2, AF_INET6=30 (not 10 as on Linux)
-const AF_INET: u16 = 2;
-const AF_INET6: u16 = 30;
+const AF_INET: u8 = 2;
+const AF_INET6: u8 = 30;
 const IFF_UP: u32 = 1;
 const IFF_LOOPBACK: u32 = 8;
